+++ app/controllers/api/v1/characters_controller.rb
@@ -4,7 +4,7 @@ module Api
   module V1
     class CharactersController < BaseController
       before_action :authenticate_player, except: :show
-      before_action :set_character, only: %i[show update destroy duplicate]
+      before_action :set_character, only: %i[show update destroy duplicate change_type]
 
       def show
         authorize @character
@@ -54,6 +54,23 @@ module Api
         else
           render json: @new_character.error.details, status: :bad_request
         end
+      end
+
+      def change_type
+        authorize @character, :update?
+        @new_type = params[:type].constantize
+
+        @new_character = @new_type.from_character!(@character)
+
+        if @new_character.save
+          render json: @new_character
+        else
+          render json: @new_character.errors.details, status: :bad_request
+        end
+
+      rescue NoMethodError, NameError
+        # TODO: Actually handle the error here
+        render json: {}, status: :bad_request
       end
 
       private